Add Misen readme file, update gitignore, and modify launch and settings files

+++ src/planetary_module/ros/package/maid_robot_system/nodes/head_unit/head_control_node/inc/models/model_structure.hpp
@@ -90,7 +90,7 @@ public:
     int lip_min = 0;
     int lip_max = 100;
 
-    int tiredness = 25;
+    double tiredness = 25.0;
 
     bool priority_to_the_right = true;
